adding load_data and saved data methods
