+++ app.rb
@@ -1,15 +1,18 @@
-require 'json'
 require_relative 'book'
 require_relative 'person'
 require_relative 'student'
 require_relative 'teacher'
 require_relative 'rental'
+require_relative 'app_function'
+require 'json'
+require 'date'
 
 class App
+  attr_accessor :people, :book, :rentals
   def initialize
-    @people = load_from_json('people.json') || []
-    @books = load_from_json('books.json') || []
-    @rentals = load_from_json('rentals.json') || []
+    @people = []
+    @books = []
+    @rentals = []
   end
 
   def display_menu
@@ -20,7 +23,7 @@ class App
     puts '4 - Add a book'
     puts '5 - Add a rental'
     puts '6 - List all rentals for a given person id'
-    puts '7 - Save and Exit'
+    puts '7 - Exit'
   end
 
   def choose_option(option)
@@ -37,100 +40,12 @@ class App
       add_rental
     when 6
       list_all_rentals
-    when 7
-      save_data
-      'Data saved. Existing...'
     else
       'Wrong number please enter a number between 1 and 7'
     end
   end
 
-  def list_all_books
-    @books.each do |book|
-      if book.nil?
-        puts "Error: Book object is nil."
-      else
-        puts "Title: \"#{book.title}\", Author: #{book.author}"
-      end
-    end
-  end
-
-  def list_all_people
-    @people.each do |person|
-      if person.instance_of?(Student)
-        puts "[Student], ID: #{person.id}, Name: #{person.name}, age: #{person.age}"
-      else
-        puts "[Teacher], ID: #{person.id}, Name: #{person.name}, age: #{person.age}"
-      end
-    end
-  end
-
-  def create_person
-    print 'Do you want to add a student (1) or a teacher (2)? [Input the number]: '
-    student_or_teacher = gets.chomp.to_i
-    case student_or_teacher
-    when 1
-      create_student
-    when 2
-      create_teacher
-    else
-      puts 'Please enter a valid number: 1 or 2'
-    end
-  end
-
-  def create_student
-    print 'Age: '
-    age = gets.chomp
-    print 'Name: '
-    name = gets.chomp
-    puts 'Has parent permission? [Y/N]'
-    permission = gets.chomp.downcase == 'y'
-    student = Student.new(age, permission, name)
-    @people << student
-    puts 'Student created sucessfully'
-  end
-
-  def create_teacher
-    print 'Age: '
-    age = gets.chomp
-    print 'Name: '
-    name = gets.chomp
-    puts 'specialization'
-    specialization = gets.chomp
-    teacher = Teacher.new(age, specialization, name)
-    @people << teacher
-    puts 'Teacher created sucessfully'
-  end
-
-  def add_book
-    print 'Title: '
-    title = gets.chomp
-    print 'Author: '
-    author = gets.chomp
-    book = Book.new(title, author)
-    @books << book
-    puts 'Book created sucessfully'
-  end
-
-  def add_rental
-    puts 'Select a book from the following list by number'
-    @books.each_with_index do |book, index|
-      puts "#{index}) Title: \"#{book.title}\", Auhtor: #{book.author}"
-    end
-    book_idx = gets.chomp.to_i
-    puts 'Select a person from the following list by number (not id)'
-    @people.each_with_index do |person, index|
-      puts "#{index}) Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
-    end
-    person_idx = gets.chomp.to_i
-
-    print 'Date: '
-    rental_date = gets.chomp
-
-    rental = Rental.new(rental_date, @books[book_idx], @people[person_idx])
-    @rentals << rental
-    puts 'Rental created successfully'
-  end
+  include AppFunctions
 
   def list_all_rentals
     print 'ID of person: '
@@ -147,36 +62,85 @@ class App
     puts 'No rentals found for the specified person.' unless rentals_found
   end
 
-   # preserve data
-   def save_data
-    save_to_json('books.json', @books.map(&:to_json))
-    save_to_json('people.json', @people.map(&:to_json))
-    save_to_json('rentals.json', @rentals.map(&:to_json))
+  def save_data
+    save_books
+    save_people
+    save_rentals
   end
 
-  def save_to_json(filename, data)
-    File.open(filename, 'w') do |file|
-      file.puts JSON.generate(data)
+  def load_data
+    load_books
+    load_people
+    load_rentals
+  end
+
+  private
+
+  def save_books
+    File.open('books.json', 'w') do |file|
+      file.write(JSON.generate(@books))
     end
   end
 
-  # load data from json
-  def load_from_json(filename)
-    return [] unless File.exist?(filename)
+  def load_books
+    if File.exist?('books.json')
+      data = JSON.parse(File.read('books.json'))
+      @books = data.map { |book_data| Book.new(book_data['title'], book_data['author']) }
+    end
+  end
 
-    JSON.parse(File.read(filename)).map do |item|
-      item = JSON.parse(item) # Parse each item from string to hash
-      puts "DEBUG: #{item}" # Add this line to print the item
-      case item['_type']
-      when 'Book'
-        Book.new(item['title'], item['author'])
-      when 'Student'
-        Student.new(item['age'], item['parent_permission'], item['name'])
-      when 'Teacher'
-        Teacher.new(item['age'], item['specialization'], item['name'])
-      when 'Rental'
-        Rental.new(item['date'], item['book'], item['person'])
+  def save_people
+    File.open('people.json', 'w') do |file|
+      file.write(JSON.generate(@people))
+    end
+  end
+
+  def load_people
+    if File.exist?('people.json')
+      data = JSON.parse(File.read('people.json'))
+
+      @people = data.map do |person_data|
+        puts "Loading person data: #{person_data.inspect}"
+
+        if person_data['class'] == 'Student'
+          Student.new(person_data['age'], person_data['parent_permission'], person_data['name'])
+        elsif person_data['class'] == 'Teacher'
+          Teacher.new(person_data['age'], person_data['specialization'], person_data['name'])
+        end
       end
     end
   end
+
+  def save_rentals
+    File.open('rentals.json', 'w') do |file|
+      file.write(JSON.generate(@rentals))
+    end
+  end
+
+  def load_rentals
+    if File.exist?('rentals.json')
+      data = JSON.parse(File.read('rentals.json'))
+
+      @rentals = data.map do |rental_data|
+        puts "Loading rental data: #{rental_data.inspect}"
+
+        book_data = rental_data['book']
+        person_data = rental_data['person']
+
+        book = @books.find { |b| b.title == book_data['title'] && b.author == book_data['author'] }
+        person = @people.find { |p| p.id == person_data['id'] }
+
+        if person
+          rental = Rental.new(Date.parse(rental_data['date']), book, person)
+          person.rentals ||= []
+          person.rentals << rental
+          rental
+        else
+          puts "Invalid rental data (person not found): #{rental_data.inspect}"
+          nil  # Ignore invalid data
+        end
+      end.compact  # Remove nil entries
+    end
+  end
+
 end
