+++ books.json
@@ -1,1 +1,1 @@
-["{\"title\":\"y\",\"author\":\"u\"}"]
+[{"title":"yu","author":"ii"},{"title":"hr","author":"huk"}]
