+++ people.json
@@ -1,1 +1,1 @@
-["{\"id\":98,\"name\":\"r\",\"age\":\"3\",\"parent_permission\":true}"]
+[{"id":33,"name":"ty","age":"44","class":"Student"}]
