+++ person.rb
@@ -25,8 +25,13 @@ class Person < Nameable
     @rentals << rental
   end
 
-  def to_json
-    {'id' => id, 'name' => name, 'age' => age}.to_json
+  def to_json(*_args)
+    {
+      id: @id,
+      name: @name,
+      age: @age,
+      class: self.class.to_s
+    }.to_json
   end
 
   private
