+++ rentals.json
@@ -1,1 +1,1 @@
-["{\"date\":\"777\",\"book\":\"{\\\"title\\\":\\\"y\\\",\\\"author\\\":\\\"u\\\"}\",\"person\":\"{\\\"id\\\":98,\\\"name\\\":\\\"r\\\",\\\"age\\\":\\\"3\\\",\\\"parent_permission\\\":true}\"}"]
+[]
