+++ student.rb
@@ -16,8 +16,4 @@ class Student < Person
   def play_hooky
     '¯\\(ツ)/¯'
   end
-
-  def to_json
-    {'id' => id, 'name' => name, 'age' => age, 'parent_permission' => parent_permission}.to_json
-  end
 end
